pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "BEACH"       # matches your package folder name
version = "0.1.0"
description = "Beam simulation and fitting tools for HIRAX"
authors = [
  { name = "Ajith Sampath", email = "ajithsampath1997@gmail.com" }
]
dependencies = [
  "numpy",
  "scipy",
  "pyyaml",
  "matplotlib",
  "astropy",
  "pandas",
  "h5py"
]
